// source/server/src/shared/enums/paginationEnum.ts
export enum Operator {
  Equals = 'Equals',
  NotEquals = 'NotEquals',
  Greater = 'Greater',
  Less = 'Less',
  GreaterOrEquals = 'GreaterOrEquals',
  LessOrEquals = 'LessOrEquals',
  Contains = 'Contains',
}

export enum Condition {
  And = 'And',
  Or = 'Or',
}

export enum SortDirection {
  Asc = 'Asc',
  Desc = 'Desc',
}

export enum FilterDataType {
  Number = 'Number',
  String = 'String',
  Boolean = 'Boolean',
}